+++ tests/OverflowTest.php
@@ -2,13 +2,12 @@
 
 namespace CraftLogan\LaravelOverflow\Tests;
 
-use CraftLogan\LaravelOverflow\Models\TestModel;
-use CraftLogan\LaravelOverflow\Requests\OverflowFormRequest;
 use Orchestra\Testbench\TestCase;
+use CraftLogan\LaravelOverflow\Models\TestModel;
+use CraftLogan\LaravelOverflow\Requests\TestOverflowFormRequest;
 
 class OverflowTest extends TestCase
 {
-
     public function setUp(): void
     {
         parent::setUp();
@@ -17,7 +16,6 @@ class OverflowTest extends TestCase
 
         (new \CreateTestModelsTable())->up();
     }
-
 
     /** @test */
     public function it_can_add_multiple_extra_properties_to_overflow()
@@ -28,14 +26,11 @@ class OverflowTest extends TestCase
             'testing' => 'i am testing',
             'extra' => 'just some more extra stuff'
         ];
-
-        $request = $this->createRequest('get', '', '/test', ['CONTENT_TYPE' => 'application/json'], $requestParams);
-
+        $request = $this->createTestRequestWithParameters($requestParams);
         $properties = json_decode($request->overflow());
 
-        $this->assertObjectHasAttribute('testing',$properties);
+        $this->assertObjectHasAttribute('testing', $properties);
     }
-
 
     /** @test */
     public function it_can_add_extra_properties_with_create_method()
@@ -45,11 +40,10 @@ class OverflowTest extends TestCase
             'description' => 'This is when you make it to the big time',
             'testing' => true,
         ];
-        $request = $this->createRequest('get', '', '/test', ['CONTENT_TYPE' => 'application/json'], $requestParams);
+        $request = $this->createTestRequestWithParameters($requestParams);
         $testmodel = TestModel::create($request->allWithOverflow());
         $this->assertTrue($testmodel->properties()->testing);
     }
-
 
     /** @test */
     public function it_can_get_table_columns()
@@ -59,16 +53,13 @@ class OverflowTest extends TestCase
             'description' => 'This is when you make it to the big time',
             'testing' => true,
         ];
-
         $tableColumnsShouldBe = [
             'id', 'name', 'description', 'properties', 'created_at', 'updated_at'
         ];
-        $request = $this->createRequest('get', '', '/test', ['CONTENT_TYPE' => 'application/json'], $requestParams);
+        $request = $this->createTestRequestWithParameters($requestParams);
         $tableColumns = $request->getTableColumns();
-
         $this->assertEqualsCanonicalizing($tableColumnsShouldBe, $tableColumns);
     }
-
 
     /** @test */
     public function it_can_add_extra_properties_using_model_attributes()
@@ -78,45 +69,23 @@ class OverflowTest extends TestCase
             'description' => 'This is when you make it to the big time',
             'testing' => true,
         ];
-
-        $request = $this->createRequest('get', '', '/test', ['CONTENT_TYPE' => 'application/json'], $requestParams);
-
+        $request = $this->createTestRequestWithParameters($requestParams);
         $testmodel = new TestModel();
         $testmodel->name = $request->name;
         $testmodel->description = $request->description;
         $testmodel->properties = $request->overflow();
         $testmodel->save();
-
         $this->assertTrue($testmodel->properties()->testing);
     }
 
-
-    protected function createRequest(
-        $method,
-        $content,
-        $uri = '/test',
-        $server = ['CONTENT_TYPE' => 'application/json'],
-        $parameters = [],
-        $cookies = [],
-        $files = []
-    ){
-        $request = new OverflowFormRequest();
-        return $request->createFromBase(
+    protected function createTestRequestWithParameters($parameters)
+    {
+        return TestOverflowFormRequest::createFromBase(
             \Symfony\Component\HttpFoundation\Request::create(
-                $uri,
-                $method,
+                'overflowTest/',
+                'POST',
                 $parameters,
-                $cookies,
-                $files,
-                $server,
-                $content
             )
         );
     }
-
-
-
-
-
-
 }
